requirements.txt
# RAG + Knowledge Graph Master Course Requirements

# Core LLM and AI
openai>=1.10.0
anthropic>=0.8.0

# Embeddings and Vector Search
sentence-transformers>=2.2.0
chromadb>=0.4.0
faiss-cpu>=1.7.4
pinecone-client>=2.2.0

# Knowledge Graph
neo4j>=5.15.0
py2neo>=2021.2.3

# Retrieval and NLP
rank-bm25>=0.2.2
spacy>=3.7.0
nltk>=3.8.1

# LLM Frameworks
langchain>=0.1.0
llama-index>=0.9.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
pypandoc>=1.12

# Web Framework
fastapi>=0.108.0
uvicorn>=0.25.0
pydantic>=2.5.0

# Database and Caching
redis>=5.0.0
sqlalchemy>=2.0.0

# Evaluation
ragas>=0.1.0
datasets>=2.16.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pandas>=2.1.0
numpy>=1.24.0
requests>=2.31.0

# Monitoring and Logging
prometheus-client>=0.19.0

# Development
jupyter>=1.0.0
ipykernel>=6.27.0
black>=23.12.0
flake8>=7.0.0
